chore: update to latest SDK
